% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.R
\name{annoBED}
\alias{annoBED}
\title{annoBED}
\usage{
annoBED(object = NULL, gtf = NULL, assay = NULL)
}
\arguments{
\item{object}{Seurat object.}

\item{gtf}{GTF database object, generated by gtf2db().}

\item{assay}{working assay}
}
\value{
Annotated Seurat object
}
\description{
Annotate gene name for bed region. Before annotation you should run ParseBED to parse chr, start and end position from feature name.
}
