% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_anno.R
\name{annoGene}
\alias{annoGene}
\title{annoGene}
\usage{
annoGene(object = NULL, assay = NULL, gtf = NULL, gene.name = NULL)
}
\arguments{
\item{object}{Seurat object.}

\item{assay}{Work assay.}

\item{gtf}{GTF object, load by gtf2db.}

\item{gene.name}{Tag name for gene name. If not set, use rownames instead and will create gene_name tag in the meta table.}
}
\value{
Annotated Seurat object with annotations.
}
\description{
Annotate genomic locations for gene names.
}
