% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varanno.R
\name{annoVAR}
\alias{annoVAR}
\title{annoVAR}
\usage{
annoVAR(
  object = NULL,
  assay = NULL,
  gtf = NULL,
  vcf = NULL,
  tags = NULL,
  check.alt.only = FALSE
)
}
\arguments{
\item{object}{Seurat object.}

\item{assay}{Work assay.}

\item{gtf}{GTF object, load by gtf2db.}

\item{vcf}{VCF database. Should be indexed with `bcftools index` at first.}

\item{tags}{Vector of tags to annotate. Require VCF database specified, and tags should be well formated in the VCF header.}

\item{check.alt.only}{Only annotate records for alternative allele (non-ref allele). Default is FASLE.}
}
\value{
Annotated Seurat object.
}
\description{
Annotate gene region and other tags for genetic variants. VCF database can be specified, and the type of VCF tags will be consider. See VCF specification and manual page for details <https://shiquan.github.io/Yano.html>.
}
