% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_test.R
\name{AssociationTest}
\alias{AssociationTest}
\title{AssociationTest}
\usage{
AssociationTest(
  object = NULL,
  features = NULL,
  assay = NULL,
  layer = "data",
  min.cells = 10,
  test.assay = NULL,
  test.features = NULL,
  bind.assay = NULL,
  bind.feature = NULL,
  wm.name = NULL,
  perm = 100,
  seed = 999,
  cells = NULL,
  threads = 0,
  verbose = TRUE,
  debug = FALSE,
  reduction = "pca",
  dims = 1:20,
  k.param = 20,
  prune.SNN = 1/50,
  return.dist = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Vector of features.}

\item{assay}{Work assay.}

\item{min.cells}{Features detected in few than minimal number of cells will be skipped. Default is 10.}

\item{test.assay}{The test assay. This option is required.}

\item{test.features}{The test features.}

\item{bind.assay}{The binding assay. If not set will use the test.assay instead.}

\item{bind.feature}{The bind feature.}

\item{wm.name}{Weight matrix name, this matrix (graph) usually be generated by \code{\link{RunAutoCorr}}.}

\item{perm}{Permutations for evaluating mean and sd of D/L scores. Default is 100.}

\item{seed}{Seed for generate random number. Default is 999.}

\item{cells}{Calculate scores for predefined cells. Will reconstruct the SNN graph and weight matrix for these cells with 'reduction' space (usually be pca or harmony). Only weight matrix that calculated by SNN is supported if cells/idents/node is defined.}

\item{threads}{Threads. If set to 0 (default), will auto check the CPU cores and set threads = number of CPU cores -1.}

\item{debug}{Print debug message. Will auto set thread to 1. Default is FALSE.}

\item{reduction}{Dimension reduction name for constructing SNN graph and weight matrix. Default is 'pca'. This and following parameters only actived when cells is set, because need to recalculate the SNN graph for the defined cells.}

\item{dims}{Dimensions of reduction used to construct SNN graph.}

\item{k.param}{Defines k for the k-nearest neighbor algorithm. This one and following parameters will passed to Seurat::FindNeighbors.}

\item{prune.SNN}{Sets the cutoff for acceptable Jaccard index when computing the neighborhood overlap for the SNN construction. Any edges with values less than or equal to this will be set to 0 and removed from the SNN graph. Essentially sets the stringency of pruning (0 --- no pruning, 1 --- prune everything). Default is 1/50.}

\item{versbose}{Print log message. Default is TRUE.}

\item{nn.method}{Method for nearest neighbor finding. Options include: rann, annoy(default).}

\item{annoy.metric}{Distance metric for annoy. Options include: euclidean (default), cosine, manhattan, and hamming}

\item{n.trees}{More trees gives higher precision when using annoy approximate nearest neighbor search. Default is 50.}

\item{nn.eps}{Error bound when performing nearest neighbor seach using RANN; default of 0.0 implies exact nearest neighbor search}
}
\description{
Run association test for features with alternative expressed pairs. Features that have similar expression pattern with the 'test feature' and show different expression pattern with the 'binding feature' will be prioritized.
}
