% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moransi.R
\name{AutoCorrFeatures}
\alias{AutoCorrFeatures}
\title{AutoCorrFeatures}
\usage{
AutoCorrFeatures(object = NULL, assay = NULL, name = "autocorr.variable")
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Working assay.}

\item{name}{Name of the autocorrelation variable column. Default is "autocorr.variable".}
}
\description{
return spatial autocorrlated features
}
