% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomSplice.R
\name{create_splice}
\alias{create_splice}
\title{create_splice}
\usage{
create_splice(x, y, xend, height = 0, spline_shape = -0.25)
}
\arguments{
\item{x}{Start x-coordinate.}

\item{y}{Start y-coordinate.}

\item{xend}{End x-coordinate.}

\item{height}{Height of the arc.}

\item{spline_shape}{Spline shape parameter, default -0.25.}
}
\value{
A data.frame with x and y coordinates for the spline path.
}
\description{
Create a spline path for a splice junction arc.
}
