% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{DimSelector}
\alias{DimSelector}
\title{DimSelector}
\usage{
DimSelector(
  object,
  dims = c(1, 2),
  cells = NULL,
  cols = NULL,
  pt.size = NULL,
  reduction = NULL,
  group.by = NULL,
  shape.by = NULL,
  order = NULL,
  shuffle = FALSE,
  seed = 1,
  label = FALSE,
  label.size = 4,
  label.color = "black",
  label.box = FALSE,
  repel = FALSE,
  alpha = 1,
  cells.highlight = NULL,
  cols.highlight = "#DE2D26",
  sizes.highlight = 1,
  na.value = "grey50",
  raster = NULL,
  raster.dpi = c(512, 512),
  return.object = FALSE,
  plot.selected = TRUE
)
}
\description{
Select cells from dimension plot. Parameters most inhert from Seurat::DimPlot().
}
