% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{FbtPlot}
\alias{FbtPlot}
\title{FbtPlot}
\usage{
FbtPlot(
  object = NULL,
  val = NULL,
  assay = NULL,
  chr.name = "chr",
  start.name = "start",
  end.name = "end",
  col.by = NULL,
  cols = NULL,
  sel.chrs = NULL,
  pt.size = NULL,
  xlab = "Chromosome",
  ylab = expression(-log[10](p)),
  subset = NULL,
  point.label = NULL,
  label.size = 3,
  shape.by = NULL,
  chr = NULL,
  start = NULL,
  end = NULL,
  gtf = NULL,
  gene = NULL,
  upstream = 1000,
  downstream = 1000,
  print.genes = NULL,
  remove.chr = FALSE,
  layout.heights = c(3, 2)
)
}
\arguments{
\item{object}{Seurat object.}

\item{val}{Specify the name of p value. The name usually be format like bind-name.pval and bind-name.padj (BH adjusted p value). For example, if you use gene as binding feature, the name should be gene_name.pval or gene_name.padj.}

\item{assay}{Work assay.}

\item{chr.name}{The title of chromosome name in the meta table. Default is "chr".}

\item{start.name}{The title of start position name in the meta table. Default is "start".}

\item{end.name}{The title of end position name in the meta table. Default is "end".}

\item{col.by}{Color points by specify the title of values in meta table. Can be discrete or continous.}

\item{cols}{Manually specify the colors. Used with col.by.}

\item{sel.chrs}{Vector of selected chromosome names to plot. Change the order by set the level of chr names.}

\item{pt.size}{Point size.}

\item{xlab}{Label for x axis. Default is "Chromosome".}

\item{ylab}{Label for y axis. Default is "-log10p".}

\item{subset}{Rule for subsetting the meta table before plot.}

\item{point.label}{Vector of points to plot their labels.}

\item{label.size}{Size of label. Default is 3.}

\item{shape.by}{Shape points by specify the title of values in meta table. Can only be discrete.}

\item{chr}{Choromsome to zoom in. Default is NULL, no zoom in. The zoom in mode can be enabled by setting chr or gene/gtf.}

\item{start}{Start position to zoom in.}

\item{end}{End position to zoom in.}

\item{gtf}{GTF database. Load by \code{gtf2db}. Default is NULL. If specified transcirpt tracks will be plotted.}

\item{gene}{Gene name to zoom in. Should used with gtf database specified.}

\item{upstream}{Flank zoom in region with upstream. Default is 1000. Only works when zoom in mode enabled.}

\item{downstream}{Flank zoom in region with downstream. Default is 1000. Only works when zoom in mode enabled.}

\item{print.genes}{Print the gene names in the transcript tracks. Default will print all or randomly 20 genes if more than 20 genes in this region.}

\item{remove.chr}{Remove 'chr' in the chromosome names.}

\item{layout.heights}{Specify the layouts for Manhatten plot and gene tracks. Default is c(3,2).}
}
\description{
Generate Manhatten plot for log10 scaled p value for spatial dissimilarity test. X axis is the location of features. The function use ggplot2 as backend.
}
