% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{FeatureSelector}
\alias{FeatureSelector}
\title{FeatureSelector}
\usage{
FeatureSelector(
  object,
  feature,
  return.object = FALSE,
  plot.selected = TRUE,
  combine = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object.}

\item{return.object}{Return Seurat object. Default return selected cell IDs.}

\item{plot.selected}{Plot the selected cells.}

\item{...}{Parameters pass to FeaturePlot.}
}
\description{
Select cells from Feature plot. Parameters most inhert from Seurat::FeaturePlot().
}
