% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltEXP.R
\name{FindAllAltExp}
\alias{FindAllAltExp}
\title{Test alternative expression for all cell groups}
\usage{
FindAllAltExp(
  object = NULL,
  assay = NULL,
  features = NULL,
  bind.name = "gene_name",
  bind.assay = NULL,
  bind.features = NULL,
  node = NULL,
  min.cells = 10,
  return.thresh = NULL,
  mode = c(1, 2, 3),
  test.use = c("DEXSeq", "PermTest"),
  threads = 0,
  perm = 100,
  seed = 999,
  pseudo.group = 3,
  debug = FALSE
)
}
\arguments{
\item{object}{A Seurat object.}

\item{assay}{Test assay (X). Default assay will be used if not set.}

\item{features}{Candidate list to test. If not set, will use AutoCorrFeatures(object, assay = assay).}

\item{bind.name}{Title name for binding features in the meta table. Consider most users start Yano to perform alternative splicing analysis, the default bind.name set to "gene_name".}

\item{bind.assay}{Bind assay (Y). If not set, will aggregate all X values of the same block.}

\item{bind.features}{Candidate list for bind features to test. If not set, will test all covered.}

\item{node}{A node to find markers for and all its children; requires \code{\link{BuildClusterTree}} to have been run previously. Only can be used if test all groups.}

\item{min.cells}{Used to filter candiate features or binding features. Require them at least expressed in min.cells. Default is 10.}

\item{return.thresh}{Only return markers that have a p-value < return.thresh. Default is NULL.}

\item{mode}{Test mode. For mode 1, X (test feature) vs Y (binding feature). For mode 2, X vs (Y-X). For mode 3, X vs (Y+X).}

\item{threads}{Threads. For DEXSeq, threads will set to 1. For other methods, threads set to 0, which will auto check the CPU cores and set threads = number of CPU cores -1.}

\item{perm}{Permutation steps for calculate statistical of delta-ratio. Default is 100.}

\item{seed}{Seed for generate random number. Default is 999.}

\item{debug}{Print debug logs. Will auto set thread to 1. Default is FALSE.}
}
\value{
Data frame containing p values.
}
\description{
Test alternative expression for all cell groups
}
