% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltEXP.R
\name{FindAltExp}
\alias{FindAltExp}
\title{Test for alternative expression using delta-ratio}
\usage{
FindAltExp(
  object = NULL,
  cells.1 = NULL,
  cells.2 = NULL,
  ident.1 = NULL,
  ident.2 = NULL,
  assay = NULL,
  bind.name = "gene_name",
  test.use = "DEXSeq",
  bind.assay = NULL,
  features = NULL,
  bind.features = NULL,
  min.cells = 10,
  return.thresh = NULL,
  mode = c(1, 2, 3),
  threads = 0,
  perm = 100,
  seed = 999,
  pseudo.group = 3,
  debug = FALSE
)
}
\arguments{
\item{object}{A Seurat object.}

\item{cells.1}{Vector of cell names belong to group 1. Conflict with ident.1}

\item{cells.2}{Vector of cell names for comparsion. Conflict with ident.2}

\item{ident.1}{Identify class to test, if not set will compare all groups one by one}

\item{ident.2}{A second class for comparsion. If NULL (default), use all other cells for comparison.}

\item{assay}{Test assay (X). Default assay will be used if not set.}

\item{bind.name}{Title name for binding features in the meta table. Consider most users start Yano to perform alternative splicing analysis, the default bind.name set to "gene_name".}

\item{bind.assay}{Bind assay (Y). If not set, will aggregate all X values of the same block.}

\item{features}{Candidate list to test. If not set, will use AutoCorrFeatures(object, assay = assay).}

\item{bind.features}{Candidate list for bind features to test. If not set, will test all covered.}

\item{min.cells}{Used to filter candiate features or binding features. Require them at least expressed in min.cells. Default is 10.}

\item{return.thresh}{Only return markers that have a p-value < return.thresh. Default is NULL.}

\item{mode}{Test mode. For mode 1, X (test feature) vs Y (binding feature). For mode 2, X vs (Y-X). For mode 3, X vs (Y+X).}

\item{threads}{Threads. If set to 0 (default), will auto check the CPU cores and set threads = number of CPU cores -1.}

\item{perm}{Permutation steps for calculate statistical of delta-ratio. Default is 100.}

\item{seed}{Seed for generate random number. Default is 999.}

\item{debug}{Print debug logs. Will auto set thread to 1. Default is FALSE.}

\item{pesudo.group}{Aggregate counts into groups for each clusters.}
}
\description{
Using delta-ratio and permutation method to simulate p value for each event.
}
\details{
This function first aggregates all raw counts per feature for each group \eqn{a} and \eqn{b}, then calculates the \eqn{delta-ratio = X_a/Y_a - X_b/Y_b}.Therefore, this function only retrieve data from layer 'counts'. It then employs a permutation method to randomize the cells in the two groups 100 times (by default) to evaluate the mean and standard deviation of delta-ratio. A p-value is calculated using a t-test.
}
