% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/different_expression.R
\name{FindDEP}
\alias{FindDEP}
\title{FindDEP}
\usage{
FindDEP(
  object = NULL,
  ident.1 = NULL,
  ident.2 = NULL,
  cells.1 = NULL,
  cells.2 = NULL,
  reduction = "pca",
  features = NULL,
  assay = NULL,
  idents = NULL,
  node = NULL,
  min.cells = 10,
  layer = "data",
  mode = c(1, 2),
  return.thresh = 0.01,
  perm = 100,
  seed = 999,
  threads = 0,
  verbose = TRUE,
  debug = FALSE,
  dims = 1:20,
  k.param = 20,
  prune.SNN = 1/50,
  n.trees = 50,
  nn.eps = 0,
  nn.method = "annoy",
  annoy.metric = "euclidean",
  setLog = TRUE
)
}
\arguments{
\item{object}{Seurat object.}

\item{ident.1}{Identify class to test, if not set will compare all groups one by one}

\item{ident.2}{A second class for comparsion. If NULL (default), use all other cells for comparison.}

\item{cells.1}{Vector of cell names belong to group 1. Conflict with ident.1}

\item{cells.2}{Vector of cell names for comparsion. Conflict with ident.2}

\item{reduction}{Dimension reduction name for constructing SNN graph and weight matrix. Default is 'pca'. This and following parameters only actived when cells is set, because need to recalculate the SNN graph for the defined cells.}

\item{features}{Vector of features to calculate. Default is AutoCorrFeatures(object).}

\item{assay}{Work assay.}

\item{idents}{Specify idents to compare. If not set will use Idents(object).}

\item{min.cells}{Features detected in few than minimal number of cells will be skipped. Default is 10.}

\item{layer}{Data layer to use, default is 'data'.}

\item{mode}{Test mode.}

\item{return.thresh}{Only return markers that have a p-value < return.thresh. Default is 0.01.}

\item{perm}{Permutations for evaluating mean and sd of D/L scores. Default is 100.}

\item{seed}{Seed for generate random number. Default is 999.}

\item{threads}{Threads. If set to 0 (default), will auto check the CPU cores and set threads = number of CPU cores -1.}

\item{debug}{Print debug message. Will auto set thread to 1. Default is FALSE.}

\item{dims}{Dimensions of reduction used to construct SNN graph.}

\item{k.param}{Defines k for the k-nearest neighbor algorithm.}

\item{prune.SNN}{Sets the cutoff for acceptable Jaccard index when computing the neighborhood overlap for the SNN construction. Any edges with values less than or equal to this will be set to 0 and removed from the SNN graph. Essentially sets the stringency of pruning (0 --- no pruning, 1 --- prune everything). Default is 1/50.}

\item{versbose}{Print log message. Default is TRUE.}

\item{...}{Parameters pass to FindNeighbors().}
}
\description{
Find different expression pattern of same feature in different cell groups. Only used for integrated data.
}
