% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{getCores}
\alias{getCores}
\title{getCores}
\usage{
getCores(threads = 0)
}
\arguments{
\item{threads}{Requested thread count. If 0, auto-detect.}
}
\value{
Integer number of threads.
}
\description{
Get number of CPU cores for parallel processing.
}
