% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{GetWeights}
\alias{GetWeights}
\title{Calcualte cell-cell weight matrix by one of shared nearest neighbour matrix, spatial locations, cell embedding and linear trajectory.}
\usage{
GetWeights(
  snn = NULL,
  pos = NULL,
  order.cells = NULL,
  emb = NULL,
  k.nn = 20,
  prune.distance = 20,
  prune.snn = 1/10,
  diag.value = 0
)
}
\arguments{
\item{snn}{Shared nearest neighbour matrix, usually can found at object[["RNA_snn"]]}

\item{order.cells}{Predefined cell ranks, used for cell lineage analysis.}

\item{emb}{Cell dimesional space (PCA/ICA/harmony)}

\item{k.nn}{k-nearest neighbors, for emb only.}

\item{diag.value}{Diagnoal value in the weight matrix.}

\item{spatial}{Tissue coordinates.}
}
\value{
A sparse weight matrix.
}
\description{
Calcualte cell-cell weight matrix by one of shared nearest neighbour matrix, spatial locations, cell embedding and linear trajectory.
}
