% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{GetWeightsFromSpatial}
\alias{GetWeightsFromSpatial}
\title{GetWeightsFromSpatial}
\usage{
GetWeightsFromSpatial(
  object = NULL,
  diag.value = 0,
  k.param = 20,
  image = NULL,
  prune.SNN = 1/30
)
}
\arguments{
\item{object}{Seurat object.}

\item{diag.value}{Diagonal value in the weight matrix. Default is 0.}

\item{k.param}{K-nearest neighbors. Default is 20.}

\item{image}{Name of the spatial image. If NULL, uses the first image.}

\item{prune.SNN}{Jaccard index cutoff for SNN pruning. Default is 1/30.}
}
\value{
A sparse weight matrix.
}
\description{
Calculate cell-cell weight matrix from spatial coordinates.
}
