% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{ggplot_selector}
\alias{ggplot_selector}
\title{ggplot_selector}
\usage{
ggplot_selector()
}
\value{
A list with x and y coordinates and logical index of selected cells.
}
\description{
Interactive cell selection tool for ggplot2-based plots.
}
