% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glbt_small}
\alias{glbt_small}
\title{glbt_small}
\format{
A Seurat object
}
\usage{
data(glbt_small)
}
\description{
A small example Seurat object containing gene expression data
for testing and examples in the Yano package.
}
