% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf2db.R
\name{gtf2db}
\alias{gtf2db}
\title{gtf2db}
\usage{
gtf2db(filename = NULL, use_utr = FALSE)
}
\arguments{
\item{filename}{Path to gtf file.}

\item{use_utr}{Load CDS records or not. Default is FALSE.}
}
\value{
A point to GTF struct.
}
\description{
Generate GTF object from gtf file.
}
