% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{ImageDimSelector}
\alias{ImageDimSelector}
\title{ImageDimSelector}
\usage{
ImageDimSelector(
  object,
  group.by = NULL,
  return.object = FALSE,
  plot.selected = TRUE,
  combine = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object.}

\item{return.object}{Return Seurat object. Default return selected cell IDs.}

\item{plot.selected}{Plot the selected cells.}

\item{...}{Parameters pass to ImageDimPlot.}
}
\description{
Select cells from Image dimension plot. Parameters most inhert from Seurat::SpatialPlot().
}
