% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{ImputationByWeight}
\alias{ImputationByWeight}
\title{ImputationByWeight}
\usage{
ImputationByWeight(X = NULL, cells = NULL, W = NULL, filter = 0.001)
}
\arguments{
\item{X}{Expression counts.}

\item{cells}{Cells to imputate.}

\item{W}{Weight matrix.}

\item{filter}{Value below this cutoff will be filtered. Used to reduce density of matrix.}
}
\description{
ImputationByWeight
}
