% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf2db.R
\name{isGTF}
\alias{isGTF}
\title{isGTF}
\usage{
isGTF(gtf = NULL)
}
\arguments{
\item{gtf}{An object to check.}
}
\value{
Logical, TRUE if the object is a GTF database.
}
\description{
Check if an object is a GTF database.
}
