% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{LoadEPTanno}
\alias{LoadEPTanno}
\title{LoadEPTanno}
\usage{
LoadEPTanno(file = NULL, object = NULL, assay = NULL, stranded = TRUE)
}
\arguments{
\item{file}{Path to the EPT annotation file.}

\item{object}{Seurat object to annotate.}

\item{assay}{Assay name. If NULL, uses default assay.}

\item{stranded}{Whether the annotation is stranded. Default is TRUE.}
}
\value{
Annotated Seurat object.
}
\description{
Load exon-pair-trio (EPT) annotation from a file.
}
