% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{mergeMatrix,SMatrix-method}
\alias{mergeMatrix,SMatrix-method}
\title{mergeMatrix}
\usage{
\S4method{mergeMatrix}{SMatrix}(x = NULL, y = NULL, ...)
}
\arguments{
\item{x}{Matrix 1}

\item{y}{Matrix 2}

\item{...}{More matrix files.}
}
\description{
Merge multiple matrix files into one. At least two matrix files should be specified. Records with the same row name and column name will be summed up.
}
