% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{mergeMatrix}
\alias{mergeMatrix}
\alias{mergeMatrix,list-method}
\title{mergeMatrix}
\usage{
mergeMatrix(x, y, ...)

\S4method{mergeMatrix}{list}(x, y, ...)
}
\arguments{
\item{x}{Matrix 1 or a list of Matrix.}

\item{y}{Matrix 2.}

\item{...}{More matrix files.}
}
\description{
Merge multiple matrix files into one. At least two matrix files should be specified.
}
