% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Meta}
\alias{Meta}
\title{Meta}
\usage{
Meta(object, assay = NULL)
}
\arguments{
\item{object}{Seurat object.}

\item{assay}{Assay name. If NULL, uses the default assay.}
}
\value{
A data.frame of meta features.
}
\description{
Access the meta feature table of a Seurat assay.
}
