% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf2db.R
\name{notGTF}
\alias{notGTF}
\title{notGTF}
\usage{
notGTF(gtf = NULL)
}
\arguments{
\item{gtf}{An object to check.}
}
\value{
Logical, TRUE if the object is not a GTF database.
}
\description{
Check if an object is NOT a GTF database.
}
