% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ParseBED}
\alias{ParseBED}
\title{ParseBED}
\usage{
ParseBED(object = NULL, assay = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{EPT assay. Default is current actived assay.}
}
\description{
Parse chromosome, start, end, strand from the BED/EPT name. The BED name is generated by `PISA annotate`.
}
