% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ParseExonName}
\alias{ParseExonName}
\title{ParseExonName}
\usage{
ParseExonName(object = NULL, assay = NULL)
}
\arguments{
\item{object}{Seurat object.}

\item{assay}{Exon assay name. Default use current actived assay.}
}
\description{
Parse chromosome, start, end, strand, and gene name from the exon name. The exon name generated by `PISA anno` is formated like chr:start-end/[+-]/gene_name.
}
\examples{
data("glbt_small")
DefaultAssay(glbt_small) <- "exon"
# Check the meta table before parsing
head(glbt_small[['exon']][[]])

glbt_small <- ParseExonName(glbt_small)

# Now see the meta table after parsing
head(glbt_small[['exon']][[]])

}
