% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{PSIPlot}
\alias{PSIPlot}
\title{PSIPlot}
\usage{
PSIPlot(
  object = NULL,
  exon.assay = NULL,
  exclude.assay = "EXCL",
  features = NULL,
  dims = c(1, 2),
  cells = NULL,
  cols = c("lightgrey", "blue"),
  pt.size = NULL,
  alpha = 1,
  order = FALSE,
  reduction = NULL,
  shape.by = NULL,
  ncol = NULL,
  split.by = NULL,
  coord.fixed = FALSE,
  combine = TRUE,
  raster = NULL,
  raster.dpi = c(512, 512),
  pesudo.num = 1
)
}
\arguments{
\item{object}{Seurat object.}

\item{exon.assay}{Exon assay name.}

\item{exclude.assay}{Excluded exon assay name.}

\item{features}{Features to plot.}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{cells}{Vector of cells to plot (default is all cells)}

\item{cols}{Vector of colors, each color corresponds to an identity class. This may also be a single character or numeric value corresponding to a palette as specified by \code{\link[RColorBrewer]{brewer.pal.info}}. By default, ggplot2 assigns colors. We also include a number of palettes from the pals package. See \code{\link{DiscretePalette}} for details.}

\item{pt.size}{Adjust point size for plotting}

\item{order}{Boolean determing whether to plot cells in order of PSI score.}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}

\item{shape.by}{If NULL, all points are circles (default). You can specify any cell attribute (that can be pulled with FetchData) allowing for both different colors and different shapes on cells.  Only applicable if \code{raster = FALSE}.}

\item{ncol}{Number of columns to combine multiple features plots to}

\item{split.by}{A factor in object metadata to split the plot by, pass 'ident' to split by cell identity}

\item{coord.fixed}{Plot cartesian coordinates with fixed aspect ratio}

\item{combine}{Combine plots into a single \code{\link[patchwork]{patchwork}} ggplot object. If \code{FALSE}, return a list of ggplot objects.}

\item{group.by}{Name of one or more metadata columns to group (color) cells by (for example, orig.ident); pass 'ident' to group by identity class}
}
\value{
A \code{{\link{[patchwork]{patchwork}}}} ggplot object of \code{combine = TRUE}; otherwise, a list of ggplot objects
}
\description{
Plot PSI score on reduction map.
}
\details{
This function is edited from Seurat::FeaturePlot, used to visulize the PSI score on cells.
}
\concept{visualization}
