% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{QuickRecipe,Seurat-method}
\alias{QuickRecipe,Seurat-method}
\title{QuickRecipe}
\usage{
\S4method{QuickRecipe}{Seurat}(
  counts = NULL,
  min.cells = 20,
  min.features = 200,
  nvar = 3000,
  resolution = 0.5,
  assay = NULL,
  ndim = 20
)
}
\arguments{
\item{counts}{raw counts matrix or Seurat object.}

\item{min.cells}{Only compute for features in at least this many cells}

\item{min.features}{Only computer for cells contained at least this many features}

\item{nvar}{Number of high variable features selected for PCA analysis}

\item{resolution}{Value of the resolution parameter pass to FindClusters, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities}

\item{assay}{Assay name. Default is 'RNA'.}

\item{ndim}{Use top N PCs for clustering and UMAP. Default is 20.}
}
\value{
Seurat object
}
\description{
Quick clust single cell gene expression matrix by Seurat pipeline
}
