% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/preprocess.R
\name{QuickRecipe0}
\alias{QuickRecipe0}
\alias{QuickRecipe0,SMatrix-method}
\alias{QuickRecipe0,Seurat-method}
\title{QuickRecipe0}
\usage{
QuickRecipe0(
  counts = NULL,
  meta.data = NULL,
  min.cells = 20,
  min.features = 200,
  ...
)

\S4method{QuickRecipe0}{SMatrix}(
  counts = NULL,
  min.cells = 20,
  min.features = 200,
  assay = NULL,
  verbose = TRUE
)

\S4method{QuickRecipe0}{Seurat}(counts = NULL, scale.factor = 10000, assay = NULL, verbose = TRUE)
}
\arguments{
\item{counts}{Raw counts matrix or Seurat object.}

\item{meta.data}{Optional meta data.}

\item{min.cells}{Minimum cells for feature filtering.}

\item{min.features}{Minimum features for cell filtering.}

\item{...}{Additional arguments.}
}
\description{
Simplified quick clustering (no dimension reduction).
}
