% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadPISA.R
\name{ReadPISA}
\alias{ReadPISA}
\title{Read feature count matrix generated by `PISA count`.}
\usage{
ReadPISA(
  mex_dir = NULL,
  barcode.path = NULL,
  feature.path = NULL,
  matrix.path = NULL,
  peak.path = NULL,
  spliced.path = NULL,
  unspliced.path = NULL,
  spanning.path = NULL,
  antisense.path = NULL,
  use_10X = FALSE,
  spatial = FALSE
)
}
\arguments{
\item{mex_dir}{Feature count outdir generated by `PISA count`.}
}
\value{
Returns a sparse matrix of feature counts or a list of spliced, unspliced, and
        spanning reads sparse matrix.
}
\description{
This function will read Matrix Market files from a directory which generated by `PISA count`.
}
