% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadPISA.R
\name{ReadStereoseq}
\alias{ReadStereoseq}
\title{ReadStereoseq}
\usage{
ReadStereoseq(count = NULL, bin = 50)
}
\arguments{
\item{count}{Path to the count matrix directory.}

\item{bin}{Bin size for spatial binning. Default is 50.}
}
\value{
A sparse matrix of feature counts.
}
\description{
Read Stereo-seq count matrix from MEX format directory.
}
