% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{RenameVARs}
\alias{RenameVARs}
\title{RenameVARs}
\usage{
RenameVARs(counts = NULL, strategy = 1)
}
\arguments{
\item{counts}{A count matrix.}

\item{strategy}{Naming strategy (1 or 2). Default is 1.}
}
\value{
A renamed count matrix.
}
\description{
Rename genetic variant features based on naming strategy.
}
