% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{RunBlockCorr}
\alias{RunBlockCorr}
\title{RunBlockCorr}
\usage{
RunBlockCorr(
  object = NULL,
  bind.name = "gene_name",
  features = NULL,
  assay = NULL,
  wm.name = NULL,
  bind.assay = NULL,
  bind.features = NULL,
  min.cells.bind = 10,
  prefix = NULL,
  min.features.per.block = 1,
  scale.factor = 10000,
  mode = c(1, 2, 3),
  library.size = NULL,
  perm = 100,
  seed = 999,
  threads = 0,
  verbose = TRUE,
  debug = FALSE,
  force = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{bind.name}{The title of binding features in the meta table is important, as most users begin Yano to perform alternative splicing analysis. By default, the `bind.name` is set to "gene_name".}

\item{features}{Vector of features to calculate. Default is AutoCorrFeatures(object).}

\item{assay}{Work assay.}

\item{wm.name}{Weight matrix name, this matrix (graph) generated by \code{\link{RunAutoCorr}}.}

\item{bind.assay}{Name of binding assay. If the binding assay is not set, raw counts of features from the same block will be aggregated first, followed by normalization.}

\item{bind.features}{List of bind features. If set, bind features will be subset with this list.}

\item{min.cells.bind}{Binding features detected in few than minimal number of cells will be skipped. Default is 10.}

\item{prefix}{Prefix name for output scores and values. Default is same with bind.name.}

\item{scale.factor}{The scale factor used to normalize counts is set to a default of 1e4. The raw counts from "counts" layer are further normalized by sample size and scale factor for the spatial dissimilarity test.}

\item{mode}{Test mode. Mode 1: Compares X (test feature) vs Y (binding feature). In this mode and bind.assay is set, normlised counts from layer 'data' will be used for comparing.  If bind.assay is none, will use raw counts from layer 'counts'. Mode 2: Compares X vs (Y - X). Mode 3: Compares X vs (Y + X). For mode 2 and 3, only support to use raw counts from layer 'counts'.}

\item{library.size}{Library size for each cell, used for normalise counts. If not set, use colSum(counts) instead.}

\item{perm}{Permutations for evaluating mean and sd of D/L scores. Default is 100.}

\item{seed}{Seed for generate random number. Default is 999.}

\item{threads}{Threads. If set to 0 (default), will auto check the CPU cores and set threads = number of CPU cores -1.}

\item{debug}{Print debug message. Will auto set thread to 1. Default is FALSE.}

\item{force}{Force to use this function.}

\item{versbose}{Print log message. Default is TRUE.}
}
\description{
Run spatial dissimilarity test for features and their binding features in parallel.
}
