% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltEXP.R
\name{RunDEXSeq}
\alias{RunDEXSeq}
\title{Test for alternative expression with DEXSeq}
\usage{
RunDEXSeq(
  object = NULL,
  bind.name = "bind_name",
  ident.1 = NULL,
  ident.2 = NULL,
  cells.1 = NULL,
  cells.2 = NULL,
  assay = NULL,
  bind.assay = NULL,
  features = NULL,
  bind.features = NULL,
  min.pct = 0.05,
  min.pct.bind.feature = 0.05,
  return.thresh = 0.01,
  node = NULL,
  pseudo.group = 3,
  mode = 1,
  threads = 1,
  debug = FALSE
)
}
\arguments{
\item{object}{A Seurat object.}

\item{bind.name}{The title of binding name in meta table.}

\item{ident.1}{Identify class to test, if not set will compare all groups one by one}

\item{ident.2}{A second class for comparsion. If NULL (default), use all other cells for comparison.}

\item{cells.1}{Vector of cell names belong to group 1. Conflict with ident.1}

\item{cells.2}{Vector of cell names for comparsion. Conflict with ident.2}

\item{assay}{Assay for test features. Default assay will be used if not set.}

\item{bind.assay}{Assay for binding features. If not set, test features in same goup (with same bind name) will be aggreated as binding feature}

\item{features}{Candidate list to test. If not set, will use all.}

\item{bind.features}{Candidate list for binding features. If not set, will use all.}

\item{min.pct}{Only test features that are detected in a minimum fraction of min.pct cells in all cells.  Meant to speed up the function by not testing genes that are very infrequenctly expressed in all cells. Remember we are testing alternative epxression pattern here, so it is possible the test feature is not expressed in one group, therefore we are not going to check by groups. Note that min.pct is set for test feature here. But in \code{\link{RunPSI}}, the min.pct is set for binding feature. Default is 0.05.}

\item{min.pct.bind.feature}{Only test binding features that are detected in a minimum fraction of min.pct.bind.feature in either of the two populations. Meant to speed up the function by not testing genes that are very infrequenctly expressed in both groups. Default is 0.05.}

\item{return.thresh}{Only return markers that have a p-value < return.thresh.}

\item{node}{A node to find markers for and all its children; requires \code{\link{BuildClusterTree}} to have been run previously. Only can be used if test all groups.}

\item{pseudo.group}{Aggregate single cells into pseudo groups, because DEXSeq is designed for bulk RNA-seq. At least 3 cells are required for each group. Default is 3.}

\item{mode}{Test mode, default is 1. See online manual for the difference between modes. <https://shiquan.github.io/Yano.html>}

\item{threads}{Threads passed to DEXSeq. Default is 1.}
}
\value{
Data frame containing p values and pct for test features and their binding features.
}
\description{
This will test the test features and binding features are different expressed between groups with a generalised linear model. See DEXSeq and DESeq2 paper for details.
}
\examples{
data("neuron_small")
alt.exon <- RunDEXSeq(object = neuron_small, assay = "flatten", bind.assay = "RNA", bind.name = "gene_name")
head(alt.exon)

}
