% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunLeeScore.R
\name{RunLeeScore}
\alias{RunLeeScore}
\title{Perform Lee score test for features and test features in parallel. This function used to select co-expressed features on single cell space.}
\usage{
RunLeeScore(
  object = NULL,
  features = NULL,
  assay = NULL,
  test.features = NULL,
  test.assay = NULL,
  layer = "data",
  weight.matrix.name = "WeightMatrix",
  perm = 100,
  threads = 0,
  seed.use = 999
)
}
\arguments{
\item{object}{Seurat object.}

\item{features}{Vector of features to calculate.}

\item{assay}{Name of assay for features. If features not found in default assay, will automatically check other assays.}

\item{test.features}{Vector of features to compare. Default is AutoCorrFeatures(object, assay = test.assay).}

\item{test.assay}{Name of assay for test.features. Default is current actived assay.}

\item{layer}{Which layer to extract data. Default is 'data' layer. So will check if the assay and test.assay are normlised first.}

\item{weight.matrix.name}{Weight matrix name, this matrix (graph) generated by \code{\link{RunAutoCorr}}.}

\item{perm}{Permutations for evaluating mean and sd of L score. Default is 100.}

\item{threads}{Threads. Default is 0, will auto check the CPU cores and set threads = number of CPU cores -1.}
}
\description{
Perform Lee score test for features and test features in parallel. This function used to select co-expressed features on single cell space.
}
