% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltEXP.R
\name{RunPSI}
\alias{RunPSI}
\title{Test for alternative expression using Percent-Spliced In}
\usage{
RunPSI(
  object = NULL,
  ident.1 = NULL,
  ident.2 = NULL,
  cells.1 = NULL,
  cells.2 = NULL,
  exon.assay = NULL,
  exclude.assay = NULL,
  features = NULL,
  genes = NULL,
  gene.name = "gene_name",
  exon.name = "exon_name",
  min.pct = 0.05,
  min.pct.exon = 0.05,
  return.thresh = NULL,
  node = NULL,
  perm = 100,
  seed = 999,
  threads = 0,
  debug = FALSE
)
}
\arguments{
\item{object}{A Seurat object.}

\item{ident.1}{Identify class to test, if not set will compare all groups one by one}

\item{ident.2}{A second class for comparsion. If NULL (default), use all other cells for comparison.}

\item{cells.1}{Vector of cell names belong to group 1. Conflict with ident.1}

\item{cells.2}{Vector of cell names for comparsion. Conflict with ident.2}

\item{exon.assay}{Assay for exons. Default assay will be used if not set.}

\item{exclude.assay}{Assay for exon exclude reads. This assay is mandantory to calculate PSI. The count matrix can be calculated by `PISA anno -exon -psi in.bam`.}

\item{features}{Candidate list to test. If not set, will use all.}

\item{genes}{Candidate list for genes to test. If not set, will test all covered.}

\item{gene.name}{Gene name in the title of meta table for exon assay. Default is "gene_name". Used if genes list is set.}

\item{exon.name}{Exon name in the title of meta table for exon assay. Default is "exon_name". The exon name should be the same in both exon and exclude assay. If not the exon.name need be predefined in the meta table.}

\item{min.pct}{Excluded + exon reads that are detected in a minimum fraction of min.pct in either of the two populations. Meant to speed up the function by not testing genes that are very infrequenctly expressed in both groups. Note that the min.pct is for binding feature. Default is 0.05.}

\item{min.pct.exon}{Only test features that are detected in a minimum fraction of min.pct.exon cells in all cells. Meant to speed up the function by not testing genes that are very infrequenctly expressed in all cells. Because we are testing alternative epxression pattern here, so it is possible the exon is not expressed in one group, therefore we are not going to check by groups. Default is 0.05.}

\item{return.thresh}{Only return markers that have a p-value < return.thresh. Default is NULL.}

\item{node}{A node to find markers for and all its children; requires \code{\link{BuildClusterTree}} to have been run previously. Only can be used if test all groups.}

\item{perm}{Permutation steps for calculate statistical of delta-PSI. Default is 100.}

\item{seed}{Seed for generate random number. Default is 999.}

\item{threads}{Threads. If set to 0 (default), will auto check the CPU cores and set threads = number of CPU cores -1.}
}
\value{
Data frame containing p values and pct for exons and their genes.
}
\description{
Using \eqn{Delta PSI} and permutation method to simulate p value for each event.
}
\details{
This function calculates the \eqn{Delta PSI = PSI_a - PSI-b} for groups \eqn{a} and \eqn{b}. It then employs a permutation method to randomize the cells in the two groups 100 times (by default) to evaluate the mean and standard deviation of delta-PSI. A p-value is calculated using a t-test. PSI stands for Percent-Spliced In and is calculated by the formula \eqn{PSI = EXON/(EXON+EXCL)}, where EXON represents reads overlapping this exon, and EXCL represents excluded reads, which are junction reads that skip this exon.
}
\examples{
data("neuron_small")
alt.exon <- RunPSI(object = neuron_small, exon.assay = "exon", exclude.assay = "excl")
head(alt.exon)

}
