% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{SDTdemo}
\alias{SDTdemo}
\title{SDTdemo}
\usage{
SDTdemo(
  object = NULL,
  bind.name = NULL,
  bind.assay = NULL,
  assay = NULL,
  mode = c(1, 2, 3),
  perm = 100,
  feature = NULL,
  library.size = NULL,
  method = c("D1", "D2")
)
}
\arguments{
\item{object}{Seurat object.}

\item{bind.name}{Name of binding feature column in meta data.}

\item{bind.assay}{Binding assay name.}

\item{assay}{Test assay name.}

\item{mode}{Test mode: 1, 2, or 3.}

\item{perm}{Number of permutations. Default is 100.}

\item{feature}{Feature to test.}

\item{library.size}{Library size vector.}

\item{method}{Method: "D1" or "D2".}
}
\value{
Test results.
}
\description{
Demo function for spatial dissimilarity testing.
}
