% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{SetAutoCorrFeatures}
\alias{SetAutoCorrFeatures}
\title{Set autocorrection by rank and/or score}
\usage{
SetAutoCorrFeatures(
  object = NULL,
  moransi.min = 0,
  assay = DefaultAssay(object),
  p.cutoff = 0.05
)
}
\description{
Set autocorrection by rank and/or score
}
