% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{SpatialConcaveHull}
\alias{SpatialConcaveHull}
\title{SpatialConcaveHull}
\usage{
SpatialConcaveHull(
  object,
  group.by = NULL,
  return.object = FALSE,
  plot.selected = TRUE,
  knn = 3,
  combine = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object.}

\item{return.object}{Return Seurat object. Default return selected cell IDs.}

\item{plot.selected}{Plot the selected cells.}

\item{knn}{Number of nearest neighbors for concave hull. Default is 3.}

\item{...}{Parameters pass to DimPlot()}
}
\description{
Select cells from spatial plot using concave hull lasso.
}
