% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{spatialDistTest}
\alias{spatialDistTest}
\title{spatialDistTest}
\usage{
spatialDistTest(coord = NULL, n = 8)
}
\arguments{
\item{coord}{A matrix of spatial coordinates.}

\item{n}{Number of nearest neighbors. Default is 8.}
}
\value{
A matrix of distances.
}
\description{
Compute minimum distances to N nearest neighbors from spatial coordinates.
}
