% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{SpatialSelector}
\alias{SpatialSelector}
\title{SpatialSelector}
\usage{
SpatialSelector(
  object,
  group.by = NULL,
  return.object = FALSE,
  plot.selected = TRUE,
  ...
)
}
\arguments{
\item{object}{Seurat object.}

\item{return.object}{Return Seurat object. Default return selected cell IDs.}

\item{plot.selected}{Plot the selected cells.}

\item{...}{Parameters pass to SpatialPlot.}
}
\description{
Select cells from Spatial plot. Parameters most inhert from Seurat::SpatialPlot().
}
