% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomSplice.R
\docType{data}
\name{StatSplice}
\alias{StatSplice}
\title{StatSplice}
\format{
An object of class \code{StatSplice} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatSplice
}
\description{
A ggplot2 Stat for computing splice junction paths.
}
\keyword{datasets}
