% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{TrackPlot}
\alias{TrackPlot}
\title{TrackPlot}
\usage{
TrackPlot(
  bamfile = NULL,
  chr = NULL,
  start = NULL,
  end = NULL,
  gene = NULL,
  strand = c("both", "forward", "reverse", "ignore"),
  split.bc = FALSE,
  bin = 1000,
  cell.tag = "CB",
  umi.tag = "UB",
  gtf = NULL,
  max.depth = 0,
  group.title.size = rel(2),
  cell.group = NULL,
  display.genes = NULL,
  meta.features = NULL,
  log.scaled = FALSE,
  upstream = 1000,
  downstream = 1000,
  fragfile = NULL,
  atac.log.scaled = FALSE,
  atac.max.depth = 0,
  col.by = NULL,
  layout.heights = c(1, 10, 2),
  highlights = NULL,
  junc = FALSE,
  junc.min.depth = 5
)
}
\arguments{
\item{bamfile}{A path to a BAM file or a list to BAM files. BAM file(s) should be indexed.}

\item{chr}{Chromosome name.}

\item{start}{Start position.}

\item{end}{End position.}

\item{strand}{Plot reads on strand. Default is bot strands. Can be one of c("both", "forward", "reverse", "ignore"). When set to ignore, read strand information will be discarded.}

\item{split.bc}{Split coverage by barcode. Default is FALSE, bulk mode.}

\item{bin}{Divide plot region into bins. Save plot time. Default is 1000.}

\item{cell.tag}{Tag for cell barcode in the BAM. Default is "CB".}

\item{umi.tag}{Tag for UMI in the BAM. Default is "UB".}

\item{gtf}{GTF database, load by gtf2db.}

\item{max.depth}{Max depth capped to plot. Default is 0, no capping.}

\item{group.title.size}{Font size for track group titles. Default is rel(2).}

\item{cell.group}{Vector or list of cell group IDs. Name for the ID is the cell name. If bamfile is a list, the cell.group can also be a list with the same length of bamfile list, and binding cell.group to bam file by the name or order in both lists. See manual online for real cases. <https://shiquan.github.io/Yano.html>}

\item{display.genes}{Vector of genes to plot in the target region. Other genes in this region will not print in transcript track plot.}

\item{meta.features}{Meta table for features. If set the regions will be also plot on top of track plot. Meta table can be accessed by object[[assay]][[]].}

\item{log.scaled}{Log scaled the coverage depth per group. Only used if the depth is super high. Disabled in default.}

\item{upstream}{Flank the target region with upstream to plot. Default is 1000.}

\item{downstream}{Flank the target region with downstream to plot. Default is 1000.}

\item{fragfile}{Fragment file for ATAC data.}

\item{atac.log.scaled}{Log scaled the coverage depth per group for ATAC tracks.}

\item{atac.max.depth}{Capped depth to plot for ATAC tracks.}

\item{col.by}{Color bed regions by value. The specified name should be a colname in meta table. Only support discrete value.}

\item{layout.heights}{Layout for track plots. Default is c(10,2) or c(1,10,2) if meta.features specified or c(1,10,10,2) if fragment file also specified.}

\item{highlights}{A region of a list of regions to hightlight. The region is format as c(start,end).}

\item{junc}{Also plot the junction reads.}

\item{junc.min.depth}{Filter out junctions if low than this cutoff. This parameter used to remove noise background. Default is 5.}
}
\description{
Plot read/UMI coverage and transcript tracks from BAM(s).
}
