% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varanno.R
\name{varanno}
\alias{varanno}
\title{varanno}
\usage{
varanno(
  chr = NULL,
  start = NULL,
  end = NULL,
  ref = NULL,
  alt = NULL,
  strand = NULL,
  vcf = NULL,
  tags = NULL,
  check.alt.only = FALSE
)
}
\arguments{
\item{chr}{Vector of chromosome names.}

\item{start}{Vector of start positions.}

\item{end}{Vector of end positions.}

\item{ref}{Vector of reference alleles.}

\item{alt}{Vector of alternative alleles.}

\item{strand}{Vector of strands.}

\item{vcf}{VCF database. Should be indexed with `bcftools index` at first.}

\item{tags}{Vector of tags to annotate. Require VCF database specified, and tags should be well formated in the VCF header.}

\item{check.alt.only}{Only annotate records for alternative allele (non-ref allele). Default is FASLE.}
}
\description{
Annotate genetic variants.
}
