% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{WeightLineage}
\alias{WeightLineage}
\title{Build cell-cell weight matrix on cell lineage.}
\usage{
WeightLineage(order.cells = NULL, k.nn = 9, self.weight = 0, scale = FALSE)
}
\arguments{
\item{order.cells}{Predefined cell ranks, used for cell lineage analysis.}

\item{k.nn}{k-nearest neighbors.}

\item{self.weight}{Diagnoal value in the weight matrix.}

\item{scale}{Scale the weight values by row.}
}
\value{
A sprase weight matrix.
}
\description{
Build cell-cell weight matrix on cell lineage.
}
