% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varanno.R
\name{annoVAR}
\alias{annoVAR}
\title{annoVAR}
\usage{
annoVAR(
  object = NULL,
  assay = NULL,
  gtf = NULL,
  vcf = NULL,
  tags = NULL,
  check.alt.only = FALSE,
  chr = "chr",
  fasta = NULL
)
}
\arguments{
\item{object}{Seurat object.}

\item{assay}{Work assay.}

\item{gtf}{GTF object, load by gtf2db.}

\item{vcf}{VCF database. Should be indexed with `bcftools index` at first.}

\item{tags}{Vector of tags to annotate. Require VCF database specified, and tags should be well formated in the VCF header. See VCF specification (https://samtools.github.io/hts-specs/VCFv4.2.pdf) for details.}

\item{check.alt.only}{Only annotate records for alternative allele (non-ref allele). Default is FASLE.}

\item{chr}{Colname name for chromsome, default is "chr".}

\item{fasta}{Genome reference in FATSA format, should be indexed with `samtools faidx` first.}
}
\value{
Annotated Seurat object.
}
\description{
Annotate genetic variants with preload GTF and/or VCF databases. Will generate gene region and other tags.
}
