% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moransi.R
\name{AutoCorrFeatures}
\alias{AutoCorrFeatures}
\title{AutoCorrFeatures}
\usage{
AutoCorrFeatures(object = NULL, assay = NULL)
}
\arguments{
\item{assay}{Working assay.}

\item{obejct}{Seurat object}
}
\description{
return spatial autocorrlated features
}
