% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{buildSNN}
\alias{buildSNN}
\title{buildSNN}
\usage{
buildSNN(knn, prune.SNN = 1/30, diag.value = 1)
}
\arguments{
\item{prune.SNN}{Sets the cutoff for acceptable Jaccard index when computing the neighborhood overlap for the SNN construction. Default is 1/30.}
}
\description{
Convert a knn list into a shared nearest neighbors adjacency matrix.
}
