% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{GetWeights}
\alias{GetWeights}
\title{Calcualte cell-cell weight matrix by feature space (usually PCA) or spatial locations.}
\usage{
GetWeights(
  object = NULL,
  reduction = "pca",
  dims = NULL,
  k.nn = 9,
  spatial = FALSE,
  self.weight = 0,
  scale = FALSE,
  order.cells = NULL,
  cells = NULL
)
}
\arguments{
\item{object}{Seurat object.}

\item{reduction}{Which dimesional reduction (PCA/ICA/harmony/spatial) to use for the weight matrix. Default is "pca".}

\item{dims}{Which dimensions to use as input features.}

\item{k.nn}{k-nearest neighbors.}

\item{spatial}{Use tissue coordinates instead of pca. Require input object is spatially resolved data.}

\item{self.weight}{Diagnoal value in the weight matrix.}

\item{scale}{Scale the weight values by row.}

\item{order.cells}{Predefined cell ranks, used for cell lineage analysis.}

\item{cells}{If set, run weight matrix on these cells only.}
}
\value{
A sparse weight matrix.
}
\description{
Calcualte cell-cell weight matrix by feature space (usually PCA) or spatial locations.
}
