% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ParseBED}
\alias{ParseBED}
\title{ParseBED}
\usage{
ParseBED(object = NULL, assay = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{EPT assay. Default is current actived assay.}
}
\description{
Parse chromosome, start, end, strand from the BED/EPT name. The BED name is generated by `PISA annotate`.
}
\examples{
data("neuron_small")
DefaultAssay(neuron_small) <- "EPT"
# Check the meta table of features before parsing
head(neuron_small[['EPT']][[]])
neuron_small <- ParseVarName(neuron_small)

# Now see the meta table after parsing
head(neuron_small[['EPT']][[]])
}
