% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{QuickRecipe,Seurat-method}
\alias{QuickRecipe,Seurat-method}
\title{Quick clust single cell gene expression matrix by Seurat pipeline}
\usage{
\S4method{QuickRecipe}{Seurat}(
  counts = NULL,
  meta.data = NULL,
  min.cells = 20,
  min.features = 200,
  nvar = 3000,
  resolution = 0.5,
  assay = NULL,
  ndim = 20,
  ...
)
}
\description{
Quick clust single cell gene expression matrix by Seurat pipeline
}
