% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadPISA.R
\name{ReadPISA}
\alias{ReadPISA}
\title{ReadPISA}
\usage{
ReadPISA(
  mex_dir = NULL,
  prefix = NULL,
  cells = NULL,
  barcode.path = NULL,
  feature.path = NULL,
  matrix.path = NULL,
  peak.path = NULL,
  spliced.path = NULL,
  unspliced.path = NULL,
  spanning.path = NULL,
  use_10X = FALSE,
  spatial = FALSE
)
}
\arguments{
\item{mex_dir}{Feature count outdir generated by `PISA count`.}

\item{prefix}{Prefix added for cell names.}

\item{cells}{Vector for candidate cells. If not set, will read all cells.}

\item{barcode.path}{Manually specify barcode path in MEX directory.}

\item{feature.path}{Manually specify feature path in MEX directory.}

\item{matrix.path}{Manually specify matrix path in MEX directory.}

\item{peak.path}{Manually specific peak path (if feature is a region) in MEX directory.}

\item{spliced.path}{Manually specific spliced matrix path in MEX directory. The spliced, unspliced and spanning read count matrix generated by `PISA count -velo`.}

\item{unspliced.path}{Manually specific unspliced matrix path.}

\item{spanning.path}{Manually specific spanning matrix path.}

\item{use_10X}{Set TRUE if feature name in the second column of feature file.}

\item{spatial}{Set TRUE if barcode is coordinate of bin. Will combine x and y coordinate into bin name "x_y".}
}
\value{
Returns a sparse matrix of feature counts or a list of spliced, unspliced, and spanning reads sparse matrix.
}
\description{
Read feature count matrix generated by `PISA count`.
}
