% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moransi.R
\name{RunAutoCorr}
\alias{RunAutoCorr}
\title{RunAutoCorr}
\usage{
RunAutoCorr(
  object = NULL,
  assay = NULL,
  layer = "data",
  reduction = "pca",
  dims = 1:20,
  k.param = 20,
  prune.SNN = 1/50,
  nn.method = "annoy",
  n.trees = 50,
  annoy.metric = "euclidean",
  nn.eps = 0,
  l2.norm = FALSE,
  cells = NULL,
  min.cells = 10,
  snn.name = NULL,
  spatial = FALSE,
  order.cells = NULL,
  weight.method = c("dist", "average"),
  prune.distance = -1,
  features = NULL,
  wm.name = NULL,
  prefix = "moransi",
  threads = 0,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Working assay}

\item{layer}{Input data layer, usually be 'data'.}

\item{reduction}{Cell space used to calculate SNN graph, default is 'pca'.}

\item{dims}{Dimensions of reduction used to calculate SNN graph.}

\item{k.param}{Defines k for K-nearest neighbor algorithm}

\item{prune.SNN}{This parameter sets the cutoff for the acceptable Jaccard index when computing neighborhood overlap during SNN (Shared Nearest Neighbor) construction. It is passed to Seurat::FindNeighbors. Any edges with Jaccard index values less than or equal to this cutoff will be set to 0 and removed from the SNN graph, effectively controlling the stringency of pruning (with 0 meaning no pruning and 1 meaning everything is pruned). The default value is 1/50, which differs from Seurat's default setting. This is because Seurat’s default is OK for cell clustering but may cause the loss of many sparse features in large cell populations during spatial dissimilarity test. Setting the cutoff to a smaller value can capture more features, but it will also increase computational time.}

\item{nn.method}{nn.method passed to Seurat::FindNeighbors, default is "euclidean".}

\item{n.trees}{n.trees passed to Seurat::FindNeighbors, default is 50.}

\item{annoy.metric}{annoy.metric passed to Seurat::FindNeighbors, default is "annoy".}

\item{nn.eps}{nn.eps passed to Seurat::FindNeighbors, default is 0}

\item{l2.norm}{L2 normalization. Default is FALSE.}

\item{cells}{Cells used for calculate weight matrix. Used with snn graph. In default will use all cells.}

\item{min.cells}{If a feature can be detect in few than min.cells, will skip to save time. Default is 10.}

\item{snn.name}{name of SNN space. If spatial=FALSE and order.cells = NULL, default snn.name will set to 'RNA_snn'. Use SNN space to calculate the cell-cell weight martix.}

\item{spatial}{Use spatial coordinate instead of SNN space and linear trajectory to calculate the cell-cell weight matrix.}

\item{order.cells}{For linear trajetory, input ordered cell names to calculate the cell-cell distance weight matrix. Conflict with sptaial=TRUE and snn.name != NULL.}

\item{weight.method}{Weight method for distance, default 1/dist^2. Also support average, use mean weight value for nearby cells.}

\item{prune.distance}{Set the cutoff for neighbors for order cells and spatial coordinates. In default, 50 for order cells, 8 for spatial coordinates.}

\item{features}{List of features to test. Default is all features with that coverage >= min.cells.}

\item{wm.name}{Weight matrix/graph name in Seurat object. After this function, the graph can be visited by obj[[wm.name]]. Default name is "RNA_wm", if you change the default name, you should specific the new name in RunBlockCorr.}

\item{prefix}{Prefix for score and p value names. Default prefix is "moransi". If you change the default name, you should specific the new name in SetAutoCorrFeatures.}

\item{threads}{Threads.}

\item{verbose}{Print log message. Default is TRUE.}

\item{...}{parameters pass to GetWeightFromSpatial, so it only works if spatial is TRUE.}
}
\description{
Calculate spatial autocorrelation (Moran's I) for features in parallel.
}
