% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moransi.R
\name{RunAutoCorr}
\alias{RunAutoCorr}
\title{Calculate Moran's index for features in parallel.}
\usage{
RunAutoCorr(
  object = NULL,
  assay = NULL,
  layer = "data",
  spatial = FALSE,
  snn.name = NULL,
  order.cells = NULL,
  prune.distance = 20,
  prune.snn = 1/10,
  cells = NULL,
  min.cells = 10,
  features = NULL,
  weight.matrix.name = "WeightMatrix",
  prefix = "moransi",
  threads = 0,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{prune.distance}{Sets cutoff for distance matrix. Any edges with distance
greater than this value will be set to 0 and remove from the distance matrix.}

\item{...}{parameters pass to}
}
\description{
Calculate Moran's index for features in parallel.
}
