% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{RunAutoCorr}
\alias{RunAutoCorr}
\title{Calculate Moran's index for features in parallel.}
\usage{
RunAutoCorr(
  object = NULL,
  assay = NULL,
  slot = "data",
  spatial = FALSE,
  W = NULL,
  scale.weight = FALSE,
  reduction = "pca",
  dims = NULL,
  k.nn = 9,
  order.cells = NULL,
  cells = NULL,
  features = NULL,
  threads = 4
)
}
\description{
Calculate Moran's index for features in parallel.
}
