% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moransi.R
\name{RunAutoCorr}
\alias{RunAutoCorr}
\title{RunAutoCorr}
\usage{
RunAutoCorr(
  object = NULL,
  assay = NULL,
  layer = "data",
  reduction = "pca",
  dims = 1:20,
  k.param = 20,
  prune.SNN = 1/30,
  cells = NULL,
  min.cells = 10,
  snn.name = NULL,
  spatial = FALSE,
  order.cells = NULL,
  weight.method = c("dist", "average"),
  nn = -1,
  features = NULL,
  wm.name = NULL,
  prefix = "moransi",
  threads = 0,
  verbose = TRUE,
  name = "autocorr.variable",
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Working assay}

\item{layer}{Input data layer, usually be 'data'.}

\item{reduction}{If snn.name is not set, the cell space will be used to calculate SNN graph. Default is 'pca'.}

\item{dims}{Dimensions of reduction used to calculate SNN graph.}

\item{k.param}{Defines k for K-nearest neighbor algorithm}

\item{prune.SNN}{This parameter sets the cutoff for the acceptable Jaccard index. It is passed to Seurat::FindNeighbors. Default is 1/30.}

\item{cells}{Cells used for calculate weight matrix. Used with snn graph. In default will use all cells.}

\item{min.cells}{If a feature can be detect in few than min.cells, will skip to save time. Default is 10.}

\item{snn.name}{Name of SNN graph. Use the SNN graph to calculate the cell-cell weight martix. The SNN graph usually be generated by Seurat::FindNeighbors().}

\item{order.cells}{For linear trajetory, input ordered cell names to calculate the cell-cell distance weight matrix. Conflict with sptaial=TRUE and snn.name != NULL.}

\item{weight.method}{Weight method for distance, default 1/dist^2. Also support average, use mean weight value for nearby cells.}

\item{nn}{Set the cutoff for nearest neighbors for order cells and spatial coordinates. In default, 50 for order cells, 8 for spatial coordinates.}

\item{features}{List of features to test. Default is all features with that coverage >= min.cells.}

\item{wm.name}{Weight matrix/graph name in Seurat object. After this function, the graph can be visited by obj[[wm.name]]. Default name is "$reduction$_wm" for reduction, "trajectory_wm" for order.cells, and "spatial_wm" for spatial coordinate.}

\item{prefix}{Prefix for score and p value names. Default prefix is "moransi". If you change the default name, you should specific the new name in SetAutoCorrFeatures.}

\item{threads}{Threads.}

\item{verbose}{Print log message. Default is TRUE.}

\item{name}{Title name to label autocorrelation features. Default is 'autocorr.variable'.}

\item{...}{parameters pass to FindNeighbors.}

\item{image}{Name of 'SpatialImage' object to get coordinates for. If set, use spatial coordinate to calculate the cell-cell weight matrix.}
}
\description{
Calculate spatial autocorrelation (Moran's I) for features in parallel. Autocorrelated features are labeled with SetAutoCorrFeatures() automatically.
}
