% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{RunBlockCorr}
\alias{RunBlockCorr}
\title{RunBlockCorr}
\usage{
RunBlockCorr(
  object = NULL,
  bind.name = "gene_name",
  features = NULL,
  assay = NULL,
  min.cells = 10,
  bind.assay = NULL,
  bind.features = NULL,
  min.cells.bind = 10,
  prefix = NULL,
  subset = NULL,
  min.features.per.block = 1,
  scale.factor = 10000,
  mode = c(1, 2, 3),
  method = c("D", "D2", "Lee"),
  library.size = NULL,
  wm.name = NULL,
  perm = 100,
  seed = 999,
  threads = 0,
  verbose = TRUE,
  debug = FALSE,
  cells = NULL,
  idents = NULL,
  node = NULL,
  reduction = "pca",
  dims = 1:10,
  k.param = 20,
  prune.SNN = 1/50,
  n.trees = 50,
  nn.eps = 0,
  nn.method = "annoy",
  annoy.metric = "euclidean"
)
}
\arguments{
\item{object}{Seurat object}

\item{bind.name}{The title of binding features in the meta table is important, as most users begin Yano to perform alternative splicing analysis. By default, the `bind.name` is set to "gene_name".}

\item{features}{Vector of features to calculate. Default is AutoCorrFeatures(object).}

\item{assay}{Work assay.}

\item{min.cells}{Features detected in few than minimal number of cells will be skipped. Default is 10.}

\item{bind.assay}{Name of binding assay. If the binding assay is not set, raw counts of features from the same block will be aggregated first, followed by normalization.}

\item{bind.features}{List of bind features. If set, bind features will be subset with this list.}

\item{min.cells.bind}{Binding features detected in few than minimal number of cells will be skipped. Default is 10.}

\item{prefix}{Prefix name for output scores and values. Default is same with bind.name.}

\item{subset}{Rules for subsetting meta table before selecting features to perform test.}

\item{scale.factor}{The scale factor used to normalize counts is set to a default of 1e4. The raw counts from "counts" layer are further normalized by sample size and scale factor for the spatial dissimilarity test.}

\item{mode}{Test mode. Mode 1: Compares X (test feature) vs Y (binding feature). In this mode and bind.assay is set, normlised counts from layer 'data' will be used for comparing.  If bind.assay is none, will use raw counts from layer 'counts'. Mode 2: Compares X vs (Y - X). Mode 3: Compares X vs (Y + X). For mode 2 and 3, only support to use raw counts from layer 'counts'.}

\item{method}{Method to use. Support D, D2 and Lee. \eqn{D = sqrt(L_x)(1-r_xy)}. \eqn{D2 = sqrt(L_y)(1-r_xy)}. Lee for Lee's Score. In default use D method, see the manual for details.}

\item{library.size}{Library size for each cell, used for normalise counts. If not set, use colSum(counts) instead.}

\item{wm.name}{Weight matrix name, this matrix (graph) generated by \code{\link{RunAutoCorr}}.}

\item{perm}{Permutations for evaluating mean and sd of D/L scores. Default is 100.}

\item{seed}{Seed for generate random number. Default is 999.}

\item{threads}{Threads. If set to 0 (default), will auto check the CPU cores and set threads = number of CPU cores -1.}

\item{debug}{Print debug message. Will auto set thread to 1. Default is FALSE.}

\item{cells}{Calculate scores for predefined cells. Will reconstruct the SNN graph and weight matrix for these cells with 'reduction' space (usually be pca or harmony). Only weight matrix that calculated by SNN is supported if cells/idents/node is defined.}

\item{idents}{Calculate scores for these cell groups. The idents should be a vector of group names in Idnets(object).}

\item{node}{A node to find markers for and all its children; requires \code{\link{BuildClusterTree}} to have been run previously. Only can be used if test all groups.}

\item{reduction}{Dimension reduction name for constructing SNN graph and weight matrix. Default is 'pca'. This and following parameters only actived when cells is set, because need to recalculate the SNN graph for the defined cells.}

\item{dims}{Dimensions of reduction used to construct SNN graph.}

\item{k.param}{Defines k for the k-nearest neighbor algorithm.}

\item{prune.SNN}{Sets the cutoff for acceptable Jaccard index when computing the neighborhood overlap for the SNN construction. Any edges with values less than or equal to this will be set to 0 and removed from the SNN graph. Essentially sets the stringency of pruning (0 --- no pruning, 1 --- prune everything). Default is 1/50.}

\item{n.trees}{More trees gives higher precision when using annoy approximate nearest neighbor search. Default is 50.}

\item{nn.eps}{Error bound when performing nearest neighbor seach using RANN; default of 0.0 implies exact nearest neighbor search}

\item{nn.method}{Method for nearest neighbor finding. Options include: rann, annoy(default).}

\item{annoy.metric}{Distance metric for annoy. Options include: euclidean (default), cosine, manhattan, and hamming}

\item{versbose}{Print log message. Default is TRUE.}
}
\description{
Run spatial dissimilarity test for features and their binding features in parallel.
}
