% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moransi.R
\name{SetAutoCorrFeatures}
\alias{SetAutoCorrFeatures}
\title{Set autocorrection by rank and/or score}
\usage{
SetAutoCorrFeatures(
  object = NULL,
  moransi.min = 0,
  assay = DefaultAssay(object),
  p.cutoff = 0.01,
  prefix = "moransi",
  verbose = TRUE
)
}
\description{
Set autocorrection by rank and/or score
}
