% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moransi.R
\name{SetAutoCorrFeatures}
\alias{SetAutoCorrFeatures}
\title{Set autocorrection by rank and/or score}
\usage{
SetAutoCorrFeatures(
  object = NULL,
  moransi.min = 0,
  assay = DefaultAssay(object),
  p.thresh = 0.01,
  prefix = "moransi",
  verbose = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{moransi.min}{Minimal score for Morans I. In default is 0.}

\item{assay}{Working assay. If not set, use DefaultAssay(object).}

\item{p.thresh}{Threshold for p value. Default is 1e-2.}

\item{prefix}{Prefix name for Moran's index and p value generated by RunAutoCorr. Default is "moransi".}

\item{verbose}{Print log message.}
}
\description{
Set autocorrection by rank and/or score
}
